Create AutomatedImageSearch.py
This code utilizes Selenium 4 to automate the image search process. It seamlessly interacts with the Google Lens interface, allowing you to upload the image and retrieve the search results. Furthermore, it performs a search on the Tineye website to check if the photo is used elsewhere. Lastly, the script saves the results as an easily accessible .html file for convenient access with just one click.